Set out outline of new test structure
